backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic-settings==2.5.2
pydantic[email]==2.8.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
proxmoxer==2.0.1
requests==2.32.3
pyyaml==6.0.2
apscheduler==3.10.4
httpx==0.27.2
